complete delete user listing functionality

+++ api/controllers/listing.controller.js
@@ -40,11 +40,20 @@ export const getListingById = async (req, res, next) => {
 
 // Delete listing
 export const deleteListing = async (req, res, next) => {
-  try {
-    const listing = await Listing.findByIdAndDelete(req.params.id);
-    if (!listing) return next(errorHandler(404, "Listing not found"));
-    res.status(200).json({ success: true, message: "Listing deleted" });
-  } catch (error) {
-    next(errorHandler(500, `Failed to delete listing: ${error.message}`));
-  }
+    const listing = await Listing.findById(req.params.id);
+
+    if (!listing) {
+      return next(errorHandler(404, "Listing not found"));
+    }
+
+    if (req.user.id !== listing.userRef) {
+      return next(errorHandler(401, 'You can only delete your own listings!'));
+    }
+
+    try{
+      await Listing.findByIdAndDelete(req.params.id);
+      res.status(200).json('Listing has been deleted!');
+    } catch (error) {
+    next(error);
+    }
 };
